// frontend/api/models/ChatMessage.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ChatMessage = {
    id?: number;
    room_id?: number;
    sender_id?: number;
    sender_nickname?: string;
    content?: string;
    created_at?: string;
};